Update SwitchList sources and tests to build in Xcode 12.
Move to requiring MacOS X 10.9 as a minimum, and remove support for plain x86.

Clean up lots of warnings and deprecated APIs.

+++ src/Engine/CargoAssigner.h
@@ -39,7 +39,7 @@
 	NSObject<RandomNumberGeneratorInterface> *generator_;
 }
 - (id) initWithEntireLayout: (EntireLayout*) layout;
-- (NSArray *) cargosForToday: (int) count;
+- (NSArray *) cargosForToday: (NSUInteger) count;
 
 // For testing only.
 - (void) setRandomNumberGenerator: (NSObject<RandomNumberGeneratorInterface>*) generator;
