+++ src/Engine/LayoutController.h
@@ -51,7 +51,7 @@
 // Creates the requested number of cargos and assigns them to unassigned cars.
 // Returns dictionary mapping car type name (as string) to NSNumber showing number of cars
 // that could not be loaded.
-- (NSMutableDictionary *) createAndAssignNewCargos: (int) loadsToAdd;
+- (NSMutableDictionary *) createAndAssignNewCargos: (NSUInteger) loadsToAdd;
 
 
 // Assigns all freight cars on the layout to the trains listed, while respecting siding lengths and doors
